LPD-78127 Type the FDS client module for import-map redirection
